UNKNOWN enum is modified to have -1 value Function<Integer, Integer> is replaced with IntUnaryOperator for finding push stream-ids

+++ src/main/java/org/reaktivity/nukleus/http2/internal/routable/stream/Settings.java
@@ -26,7 +26,7 @@ public class Settings
 
     int initialWindowSize = 65_535;
 
-    long maxFrameSize = 16_384;
+    int maxFrameSize = 16_384;
 
     long maxHeaderListSize;
 }
